refactor: popup 관련 styled-componenet 분리 및 타입 분리
