+++ Client/src/components/Comment/Popup.tsx
@@ -1,14 +1,14 @@
 import Modal, { Styles } from 'react-modal';
-import { AgreeButton } from '../Commons/styled';
+import {
+  AgreeButton,
+  BtnContainer,
+  CloseBtn,
+  ClosePopupButton,
+  SubMentContainer,
+  Title,
+} from '../Commons/styled';
 import { useRecoilValue } from 'recoil';
 import { commentIdState } from '../../recoil/board/atoms';
-import {
-  Title,
-  CloseBtn,
-  SubMentContainer,
-  BtnContainer,
-  ClosePopupButton,
-} from './styled';
 import { useDeleteComment } from '../../services/board';
 
 Modal.setAppElement('#root');
