+++ Client/src/components/Comment/styled.ts
@@ -1,7 +1,6 @@
 import styled from 'styled-components';
 import { Colors } from '../../style/common.styled';
-import { AgreeButton, DisagreeButton } from '../Commons/styled';
-import { ReactComponent as CloseBtnImg } from '../../assets/icons/btn-close-gray.svg';
+import { AgreeButton } from '../Commons/styled';
 
 export const CommentContainer = styled.div`
   width: 80%;
@@ -125,44 +124,4 @@ export const CommentTextInput = styled.input`
   &:focus {
     outline: none;
   }
-`;
-
-export const CloseBtn = styled(CloseBtnImg)`
-  width: 16px;
-  height: 16px;
-  cursor: pointer;
-`;
-
-export const Title = styled.div`
-  font-family: var(--font-nanumfont);
-  font-size: 22px;
-  font-weight: 900;
-  color: ${Colors.vDeepGray};
-  padding: 0 20px;
-  position: relative;
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-`;
-
-export const SubMentContainer = styled.div`
-  margin-top: 15px;
-  width: 100%;
-  border-top: 1px solid #e5e5eb;
-  border-bottom: 1px solid #e5e5eb;
-  padding: 20px;
-  font-family: var(--font-nanumfont);
-  font-size: 17px;
-  color: ${Colors.vDeepGray};
-`;
-
-export const BtnContainer = styled.div`
-  width: 100%;
-  display: flex;
-  justify-content: flex-end;
-  padding: 20px 19px 0px 0px;
-`;
-
-export const ClosePopupButton = styled(DisagreeButton)`
-  margin-right: 15px;
 `;
