+++ Client/src/components/Commons/styled.ts
@@ -1,6 +1,7 @@
 import { Navbar } from 'react-bootstrap';
 import styled from 'styled-components';
-import { Colors } from '../../style/common.styled';
+import { Colors, StyledCommonflexCenter } from '../../style/common.styled';
+import { ReactComponent as CloseBtnImg } from '../../assets/icons/btn-close-gray.svg';
 
 const Header = styled(Navbar)`
   width: 100%;
@@ -76,10 +77,72 @@ export const DisagreeButton = styled.button<ButtonProps>`
   }
 `;
 
+export const CloseBtn = styled(CloseBtnImg)`
+  width: 16px;
+  height: 16px;
+  cursor: pointer;
+`;
+
+export const Title = styled.div`
+  font-family: var(--font-nanumfont);
+  font-size: 22px;
+  font-weight: 900;
+  color: ${Colors.vDeepGray};
+  padding: 0 20px;
+  position: relative;
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+`;
+
+export const SubMentContainer = styled.div`
+  margin-top: 15px;
+  width: 100%;
+  border-top: 1px solid #e5e5eb;
+  border-bottom: 1px solid #e5e5eb;
+  padding: 20px;
+  font-family: var(--font-nanumfont);
+  font-size: 17px;
+  color: ${Colors.vDeepGray};
+`;
+
+export const SubMentCenterContainer = styled.div`
+  margin-top: 15px;
+  width: 100%;
+  border-top: 1px solid #e5e5eb;
+  border-bottom: 1px solid #e5e5eb;
+  padding: 20px;
+  font-family: var(--font-nanumfont);
+  font-size: 17px;
+  color: ${Colors.vDeepGray};
+  flex-flow: column wrap;
+  ${StyledCommonflexCenter}
+`;
+
+export const BtnContainer = styled.div`
+  width: 100%;
+  display: flex;
+  justify-content: flex-end;
+  padding: 20px 19px 0px 0px;
+`;
+
+export const ClosePopupButton = styled(DisagreeButton)`
+  margin-right: 15px;
+`;
+
 export const S = {
   Header,
   LogoutButton,
   SpinnerWrapper,
   AgreeButton,
   DisagreeButton,
+};
+
+export const PopUp = {
+  ClosePopupButton,
+  BtnContainer,
+  SubMentCenterContainer,
+  SubMentContainer,
+  Title,
+  CloseBtn,
 };
